required-scc: don't add SA, use only the user privileges
This commit also moves the test to the scc.go file since it's
in fact testing SCC behavior, not PSa.
